notification added and availabilities controller updated

+++ internal/models/serviceModel.go
@@ -11,5 +11,5 @@ type Service struct {
 	Provider        Provider `gorm:"foreignKey:ProviderID"`
 	Description     string   `gorm:"type:text"`
 	DurationMinutes uint     `gorm:"not null;check:duration_minutes > 0"`
-	Price           float32  `gorm:"not null;check:price > 0"`
+	Price           float64  `gorm:"not null;check:price > 0"`
 }
